Pin pip for sandboxes.
Using a newer version of pip for sandboxes seems to be causing a
permission denied error when trying to to install sandbox-packages and
symmath from common/lib.  Putting this in place to work around that so
that production deploys aren't blocked on the issue.

(cherry picked from commit bdd14de365b9154c68cec940895036070c23f779)