Pin retirement service (tubular) to lilac.master by default
Previously, we pinned to an opencraft release as a quick fix,
but that quick fix is now merged to lilac upstream.
(The fix being https://github.com/edx/tubular/pull/531 )
It didn't make it to lilac.2, so we need to pin to lilac.master.

(cherry picked from commit 06fb8f4f4921601695586f080c2a6e68045c1602)